# .context/openapi.yaml
openapi: 3.0.3
info:
  title: SISE - Multimodal Retrieval & Privacy-Aware Search API
  description: >
    Production-ready contract for SISE. v1.0.0 includes:
    - flexible vector dimensions with server-side validation and explicit error codes,
    - recursive structured filters with canonical examples,
    - presigned upload flow with idempotency semantics,
    - standardized error responses and operation examples,
    - dependency-aware health probes exposing expected vector dimension.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development gateway (FastAPI)
security:
  - BearerAuth: []
paths:
...
  /auth/register: # Đăng ký người dùng mới. Đây là endpoint công khai, không yêu cầu xác thực. Người dùng cung cấp username, email và password để tạo tài khoản mới.
    post:
      operationId: registerUser
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username: { type: string }
                email: { type: string, format: email }
                password: { type: string }
            examples:
              basic:
                value:
                  username: "alice" # Đây là tên người dùng mới. Nếu username đã tồn tại, server sẽ trả về lỗi 400 Bad Request.
                  email: "alice@example.com" # Đây là email của người dùng mới. Nếu email đã tồn tại hoặc định dạng sai, server sẽ trả về lỗi 400 Bad Request.
                  password: "P@ssw0rd!" # Đây là mật khẩu của người dùng mới. Nếu mật khẩu không hợp lệ, server sẽ trả về lỗi 400 Bad Request.
      responses:
        '201': # Status 201: Created, có nghĩa là người dùng mới đã được tạo thành công. Phản hồi thường bao gồm thông tin người dùng (không bao gồm mật khẩu) và có thể kèm theo token xác thực nếu đăng ký cũng thực hiện đăng nhập.
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ, ví dụ thiếu trường bắt buộc, định dạng email sai, hoặc username đã tồn tại.
          $ref: '#/components/responses/StandardError'
        '500': # Status 500: Internal Server Error, có nghĩa là có lỗi nghiêm trọng trong quá trình xử lý yêu cầu, có thể do lỗi code hoặc sự cố môi trường.
          $ref: '#/components/responses/StandardError'
  # ------------------------------------------------------------------------------------------------------------------
  /auth/login: # Xác thực người dùng và lấy JWT token. Đây là endpoint công khai, không yêu cầu xác thực. Người dùng cung cấp username và password để nhận được access token (JWT) nếu thông tin đăng nhập hợp lệ.
    post:
      operationId: loginUser
      summary: Authenticate and obtain JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
            examples:
              basic:
                value:
                  username: "alice" # Đây là tên người dùng đã đăng ký trước đó. Nếu username không tồn tại trong hệ thống, server sẽ trả về lỗi 401 Unauthorized.
                  password: "P@ssw0rd!" # Đây là mật khẩu của người dùng. Nếu mật khẩu không đúng, server sẽ trả về lỗi 401 Unauthorized.
      responses:
        '200': # Status 200: OK, có nghĩa là xác thực thành công và access token đã được trả về trong phản hồi. Token này sẽ được sử dụng cho các yêu cầu tiếp theo đến các endpoint bảo vệ.
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401': # Status 401: Unauthorized, có nghĩa là thông tin đăng nhập không hợp lệ (username không tồn tại hoặc password sai).
          $ref: '#/components/responses/Unauthorized'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ, ví dụ thiếu trường bắt buộc hoặc định dạng sai.
          $ref: '#/components/responses/StandardError'
  /auth/me: # Retrieve the authenticated user's profile information. Endpoint này yêu cầu xác thực và cho phép user lấy thông tin profile của chính mình. Server sẽ kiểm tra JWT token và trả về thông tin người dùng hiện tại.
    get:
      operationId: getCurrentUser
      summary: Retrieve authenticated user's profile
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        '200': # Status 200: OK, có nghĩa là thông tin profile người dùng đã được truy xuất thành công.
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
...
  /search/image: # Search by image (app-level wrapper around inference + hybrid search). Endpoint này yêu cầu xác thực và cho phép user tìm kiếm ảnh tương tự bằng cách gửi một tệp ảnh.
    post:
      operationId: searchByImage
      summary: Search by image (app-level wrapper around inference + hybrid search)
      tags: [Search]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                top_k: { type: integer, default: 10 }
                metric: { $ref: '#/components/schemas/MetricType' }
                album_id: { type: integer, description: "Optional: filter results to specific album" }
      responses:
        '200': # Status 200: OK, có nghĩa là tìm kiếm đã thực hiện thành công.
          description: Search results by image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'
  /search/text: # Search by text query (app-level wrapper around text embedding + hybrid search). Endpoint này yêu cầu xác thực và cho phép user tìm kiếm ảnh bằng cách gửi một chuỗi văn bản mô tả.
    post:
      operationId: searchByText
      summary: Search by text query (app-level wrapper around text embedding + hybrid search)
      tags: [Search]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query_text]
              properties:
                query_text: { type: string }
                top_k: { type: integer, default: 10 }
                metric: { $ref: '#/components/schemas/MetricType' }
                album_id: { type: integer, description: "Optional: filter results to specific album" }
      responses:
        '200': # Status 200: OK, có nghĩa là tìm kiếm đã thực hiện thành công.
          description: Search results by text
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400': # Status 400: Bad Request, có nghĩa là dữ liệu đầu vào không hợp lệ.
          $ref: '#/components/responses/StandardError'
        '401': # Status 401: Unauthorized, có nghĩa là yêu cầu không có token hợp lệ hoặc token đã hết hạn.
          $ref: '#/components/responses/Unauthorized'

          # --- VECTOR INDEXING & SEARCH -------------------------------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    StandardError:
      description: Standard error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    PrivacyLevel: # Privacy level of media (0: Private, 1: Friends, 2: Public)
      type: integer
      enum: [0, 1, 2]
      description: "0: Private, 1: Friends, 2: Public"
    MetricType: # Similarity metric for ANN search
      type: string
      enum: [L2, IP, COSINE]
      default: COSINE

    # --- VECTOR & EMBEDDING ---------------------------------
...
    ImageMetadata: # Metadata associated with an image, used for access control and search filtering
      type: object
      properties:
        image_id: { type: string, format: uuid }
        user_id: { type: integer }
        album_id: { type: integer }
        minio_url: { type: string, format: uri }
        privacy_level: { $ref: '#/components/schemas/PrivacyLevel' }
        tags:
          type: array
          items: { type: string }
        created_at:
          type: string
          format: date-time
        index_status:
          type: string
          example: "ready"

    # --- SEARCH RESPONSE --------------------------------------
    SearchResultItem: # Individual search result item returned from hybrid search endpoint
      type: object
      properties:
        image_id: { type: string, format: uuid }
        score: { type: number }
        minio_url: { type: string, format: uri }
        metadata:
          $ref: '#/components/schemas/ImageMetadata'
    SearchResponse: # Response schema for hybrid search endpoint
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultItem'
        latency_ms:
          type: number
          description: Query latency in milliseconds
        top_k:
          type: integer

    # --- AUTH MODELS -----------------------------------------
    AuthRequest: # Request schema for authentication endpoints (register, login)
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthResponse: # Response schema for authentication endpoints (register, login)
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: "bearer" }
        expires_in: { type: integer, example: 3600 }
    User: # User model representing a registered user in the system
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
...
    Error: # Standard error response schema for all endpoints
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "ERR_VECTOR_DIM_MISMATCH"
        message:
          type: string
          example: "Expected 512 items, received 768"
        details:
          type: object
          description: Optional structured details for debugging
...